Emergency backup created
See issue 6. Backup script now copies jenkins_home to a new folder in
S3. This stops the existing backup getting corrupted if the copy is
interrupted. Once complete the new backup is moved to the usual location
on S3 - the move operation between S3 locations appears to handle
interruptions cleanly